Data hub accessiblity fixes (#1715)
* DP-26730: Add autocomplete attribute to HeaderSearch molecule

* DP-26730: Add autocomplete attribute to HeaderSearch molecule

* DP-26730: Add autocomplete attribute to HeaderSearch molecule

* DP-26747: Added disabled property for SelectBox

* DP-26734: Added visually hidden form title to SearchBannerForm

* DP-26734: Added visually hidden form title to SearchBannerForm

* DP-26734: Added visually hidden form title to SearchBannerForm

* DP-26734: Added visually hidden form title to SearchBannerForm

* DP-26734: Added visually hidden form title to SearchBannerForm

* DP-26734: Added visually hidden form title to SearchBannerForm

* DP-26746: Added clear all button context for Tags molecule

* DP-26746: Added clear all button context for Tags molecule

* DP-26746: Added clear all button context for Tags molecule

* DP-26749: After clicking a new page, allow screen reader users to go back to top

* DP-26749: After clicking a new page, allow screen reader users to go back to top

* DP-26743: Added role attribute to InputCheckbox

* DP-26743: Added role attribute to InputCheckbox

* DP-26743: Added role attribute to InputCheckbox

* DP-26752: Allows extra properties to be passed to BreadcrumbItem

* DP-26749: Added class to back to top pagination link

* linting

* DP-26749: Adds styles to back to top pagination link

* Style lint fixes

* DP-26749: Adds styles to back to top pagination link

* DP-26747: Added disabled property for SelectBox

* Added PR numbers to the change log description

* DP-26222: Fix failing backstop test for link list

* Revert "DP-26222: Fix failing backstop test for link list"

This reverts commit 8a7ad282e0aba4669d334b4b59cfbd8b40a7bfe7.

* DP-26714: Refactor GenTeaser into multiple separate components (#1717)

* DP-26714: Split GenTeaser component into a set of independent components.

Maintain backward compatibility, so that we don't have to release another major release.

* DP-26714: Update the docs & the Storybook story with the GenTeaser refactoring.

* DP-26714: Fix code style issues.

* DP-26714: Add changelog.

* DP-26714: Remove internal components - TeaserOrgs and TeaserSearch. Update the changelog.

* DP-26714: Document the legacy usage method of GenTeaser.


* DP-26745: A11y improvements to `Tags` and `ButtonTag` (#1716)

* DP-26745: Let the button tag content be a node, not just plain text.

* DP-26745: Add changelog.

* DP-26745: Render the icon with HTML rather than CSS. Add description prop.

* Hide the icon from screen readers.

* style lint
