+++ packages/assets/scss/01-atoms/_button-tag.scss
@@ -15,8 +15,7 @@
   color: $c-font-link;
   font-weight: $fonts-bold;
 
-  &:after {
-    content: "+";
+  &__icon {
     display: inline-block;
     font-size: 1.5em;
     line-height: 1.125rem;
