+++ packages/assets/scss/02-molecules/_download-link.scss
@@ -53,7 +53,7 @@
       margin-right: 0;
 
       .ma__inline-links__item:nth-last-child(2) {
-        
+
         &:after {
           visibility: hidden;
         }
