+++ packages/react/src/components/atoms/buttons/ButtonTag/index.js
@@ -7,7 +7,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const ButtonTag = ({
-  type, value, text, handleClick
+  type, value, text, description, handleClick
 }) => (
   <button
     type="button"
@@ -17,6 +17,14 @@ const ButtonTag = ({
     onClick={handleClick}
   >
     {text}
+    {!!description && (
+      <span className="visually-hidden">
+        {description}
+      </span>
+    )}
+    <span className="ma__button-tag__icon" aria-hidden="true">
+      +
+    </span>
   </button>
 );
 
@@ -25,8 +33,10 @@ ButtonTag.propTypes = {
   type: PropTypes.string.isRequired,
   /** The label text of the sort button */
   value: PropTypes.string.isRequired,
-  /** The label text of the sort button */
-  text: PropTypes.string.isRequired,
+  /** The content of the button */
+  text: PropTypes.node.isRequired,
+  /** The visually hidden description of the button */
+  description: PropTypes.node,
   /** An array of sort button objects */
   handleClick: PropTypes.func.isRequired
 };
