+++ packages/react/src/components/molecules/Tags/index.js
@@ -44,13 +44,17 @@ class Tags extends React.Component {
             type={tag.type}
             value={tag.value}
             text={tag.text}
+            description={tag.description}
             handleClick={(e) => this.handleClearThis(e)}
             /* eslint-disable-next-line react/no-array-index-key */
             key={`resultsHeading.tag.${tagIndex}`}
           />
         ))}
         { tags.length > 1 && (
-        <button type="button" className="ma__tags-clear js-results-heading-clear" onClick={() => this.handleClearAll()}>Clear all</button>
+        <button type="button" className="ma__tags-clear js-results-heading-clear" onClick={() => this.handleClearAll()}>
+          Clear all
+          {this.props.clearAllContext && (<span className="ma__visually-hidden">{` ${this.props.clearAllContext}`}</span>)}
+        </button>
         )}
       </div>
       )
@@ -61,17 +65,21 @@ class Tags extends React.Component {
 Tags.propTypes = {
   /** The tags applied to the search list <br/>
         type: The type of tag <br />
-        text: The text displayed by the tag (required) <br />
+        text: The content displayed by the tag (required) <br />
+        description: The visually hidden description of the tag <br />
         value: The value of the tag  */
   tags: PropTypes.arrayOf(PropTypes.shape({
     type: PropTypes.string,
-    text: PropTypes.string.isRequired,
+    text: PropTypes.node.isRequired,
+    description: PropTypes.node,
     value: PropTypes.string
   })),
   /** Custom onClick function that triggers when 'Clear all' button is clicked */
   onClearCallback: PropTypes.func,
   /** Custom onClick function that triggers when a tag is clicked */
-  onClearThisCallback: PropTypes.func
+  onClearThisCallback: PropTypes.func,
+  /** Clear all button context for screen readers. */
+  clearAllContext: PropTypes.string
 };
 
 export default Tags;
